refactor(deadcode): trim pairing test exports
