+++ scripts/deadcode-exports.baseline.mjs
@@ -1157,7 +1157,6 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "extensions/zalouser/src/zca-client.ts: ThreadType",
   "src/acp/control-plane/active-turns.ts: resetAcpActiveTurnsForTests",
   "src/acp/runtime/registry.ts: AcpRuntimeBackend",
-  "src/acp/runtime/session-meta.ts: resolveSessionStorePathForAcp",
   "src/agents/agent-bundle-lsp-runtime.ts: spawnLspServerProcess",
   "src/agents/agent-bundle-mcp-tools.ts: __testing",
   "src/agents/agent-bundle-mcp-tools.ts: BundleMcpToolRuntime",
@@ -1243,6 +1242,7 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/agents/cli-credentials.ts: readCodexCliCredentials (cliCredentials)",
   "src/agents/cli-credentials.ts: resetCliCredentialCachesForTest",
   "src/agents/cli-executable-identity.ts: CliExecutableFileIdentity",
+  "src/agents/cli-output.ts: CliTerminalFailure",
   "src/agents/cli-output.ts: parseCliJson",
   "src/agents/cli-output.ts: parseCliJsonl",
   "src/agents/cli-output.ts: supportsCliJsonlToolEvents",
@@ -1442,6 +1442,12 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/agents/modes/interactive/theme/theme.ts: ThemeBg",
   "src/agents/modes/interactive/theme/theme.ts: ThemeColor",
   "src/agents/openai-completions-compat.ts: resolveOpenAICompletionsCompatDefaults",
+  "src/agents/openai-completions-transport.ts: parseTransportChunkUsage",
+  "src/agents/openai-responses-transport.ts: buildOpenAIResponsesParams",
+  "src/agents/openai-responses-transport.ts: OpenAIResponsesRequestParams",
+  "src/agents/openai-responses-transport.ts: resolveAzureOpenAIApiVersion",
+  "src/agents/openai-responses-transport.ts: resolveProviderTransportTurnState",
+  "src/agents/openai-responses-transport.ts: sanitizeTransportPayloadText",
   "src/agents/openai-transport-stream.ts: __testing",
   "src/agents/openai-transport-stream.ts: testing",
   "src/agents/openclaw-tools.ts: testing",
@@ -1668,43 +1674,18 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/audit/audit-event-store.ts: auditEventStoreLimits",
   "src/audit/audit-event-store.ts: testApi",
   "src/audit/audit-event-types.ts: AgentRunAuditEventInput",
-  "src/audit/audit-event-types.ts: AuditEventKind",
-  "src/audit/audit-event-types.ts: AuditEventStatus",
-  "src/audit/audit-event-types.ts: AuditInboundMessageFailureReasonCode",
-  "src/audit/audit-event-types.ts: AuditMessageDirection",
   "src/audit/audit-event-types.ts: InboundMessageAuditEventInput",
   "src/audit/audit-event-types.ts: OutboundMessageAuditEventInput",
   "src/audit/audit-event-types.ts: OutboundMessageAuditTerminal",
-  "src/audit/audit-event-types.ts: ToolActionFinishedAuditTerminal",
   "src/audit/audit-event-writer.ts: testApi",
-  "src/audit/audit-identity.ts: AuditIdentityKey",
-  "src/audit/audit-identity.ts: AuditIdentityKind",
-  "src/audit/audit-recorder.ts: AuditEventRecorder",
   "src/audit/message-audit-events.ts: resetMessageAuditEventsForTest",
   "src/auto-reply/command-detection.runtime-types.ts: IsControlCommandMessage",
   "src/auto-reply/command-detection.runtime-types.ts: ShouldComputeCommandAuthorized",
   "src/auto-reply/commands-registry.runtime-types.ts: ShouldHandleTextCommands",
-  "src/auto-reply/dispatch.ts: DispatchInboundResult",
-  "src/auto-reply/fallback-state.ts: FallbackNoticeState",
-  "src/auto-reply/fallback-state.ts: resolveActiveFallbackState",
   "src/auto-reply/reply-payload.ts: buildPairingQrReplyChannelData",
-  "src/auto-reply/reply-payload.ts: PairingQrReplyChannelData",
-  "src/auto-reply/reply.ts: extractElevatedDirective",
-  "src/auto-reply/reply.ts: extractExecDirective",
-  "src/auto-reply/reply.ts: extractQueueDirective",
-  "src/auto-reply/reply.ts: extractReasoningDirective",
-  "src/auto-reply/reply.ts: extractReplyToTag",
-  "src/auto-reply/reply.ts: extractThinkDirective",
-  "src/auto-reply/reply.ts: extractTraceDirective",
-  "src/auto-reply/reply.ts: extractVerboseDirective",
-  "src/auto-reply/reply.ts: GetReplyOptions",
-  "src/auto-reply/reply.ts: ReplyPayload",
-  "src/auto-reply/reply/abort.ts: __testing",
   "src/auto-reply/reply/abort.ts: getAbortMemory",
   "src/auto-reply/reply/abort.ts: getAbortMemorySizeForTest",
   "src/auto-reply/reply/abort.ts: resetAbortMemoryForTest",
-  "src/auto-reply/reply/abort.ts: resolveAbortCutoffFromContext",
-  "src/auto-reply/reply/abort.ts: shouldSkipMessageByAbortCutoff",
   "src/auto-reply/reply/abort.ts: testing",
   "src/auto-reply/reply/acp-reset-target.ts: testing",
   "src/auto-reply/reply/agent-runner-cli-dispatch.ts: ReasoningProgressPayload",
@@ -1713,47 +1694,28 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/auto-reply/reply/agent-runner-execution.ts: buildContextOverflowRecoveryText",
   "src/auto-reply/reply/agent-runner-execution.ts: computeContextAwareReserveTokensFloor",
   "src/auto-reply/reply/agent-runner-execution.ts: MAX_LIVE_SWITCH_RETRIES",
-  "src/auto-reply/reply/agent-runner-memory.ts: SessionTranscriptUsageSnapshot",
   "src/auto-reply/reply/agent-runner-memory.ts: setAgentRunnerMemoryTestDeps",
   "src/auto-reply/reply/agent-runner-session-reset.ts: setAgentRunnerSessionResetTestDeps",
   "src/auto-reply/reply/agent-runner-utils.ts: buildEmbeddedRunBaseParams",
   "src/auto-reply/reply/agent-runner-utils.ts: resolveProviderScopedAuthProfile",
   "src/auto-reply/reply/block-reply-pipeline.ts: createBlockReplyPayloadKey",
-  "src/auto-reply/reply/commands-acp/shared.ts: ParsedSetCommandInput",
-  "src/auto-reply/reply/commands-acp/shared.ts: ParsedSingleValueCommandInput",
-  "src/auto-reply/reply/commands-acp/shared.ts: ParsedSpawnInput",
-  "src/auto-reply/reply/commands-acp/shared.ts: ParsedSteerInput",
   "src/auto-reply/reply/commands-diagnostics.ts: createDiagnosticsCommandHandler",
-  "src/auto-reply/reply/commands-info.ts: handleContextCommand",
-  "src/auto-reply/reply/commands-info.ts: handleWhoamiCommand",
   "src/auto-reply/reply/commands-login.ts: testing",
   "src/auto-reply/reply/commands-mcp.ts: createMcpCommandHandler",
   "src/auto-reply/reply/commands-name.ts: parseNameCommand",
   "src/auto-reply/reply/commands-setunset.ts: parseSetUnsetCommand",
   "src/auto-reply/reply/commands-setunset.ts: parseSetUnsetCommandAction",
-  "src/auto-reply/reply/commands-slash-parse.ts: ParsedSlashCommand",
-  "src/auto-reply/reply/commands-steer.runtime.ts: queueEmbeddedAgentMessage",
-  "src/auto-reply/reply/commands-subagents-dispatch.ts: COMMAND",
   "src/auto-reply/reply/commands-subagents/shared.ts: COMMAND",
   "src/auto-reply/reply/commands-tasks.ts: buildTasksReply",
   "src/auto-reply/reply/completion-delivery-policy.ts: resolveCompletionChatType",
   "src/auto-reply/reply/delivery-hints.ts: MESSAGE_TOOL_ONLY_DELIVERY_HINT",
   "src/auto-reply/reply/delivery-hints.ts: ROOM_EVENT_DELIVERY_HINT",
-  "src/auto-reply/reply/directive-handling.model-runtime.ts: ModelRuntimeDirectiveResolution",
-  "src/auto-reply/reply/directive-handling.model.ts: resolveModelSelectionFromDirective",
-  "src/auto-reply/reply/directive-handling.params.ts: HandleDirectiveOnlyCoreParams",
-  "src/auto-reply/reply/directive-handling.persist.ts: PersistedThinkingLevelRemap",
-  "src/auto-reply/reply/directives.ts: NoticeLevel",
-  "src/auto-reply/reply/dispatch-acp-delivery.ts: AcpDispatchDeliveryMeta",
-  "src/auto-reply/reply/dispatch-from-config.runtime.ts: updateSessionStoreEntry",
-  "src/auto-reply/reply/dispatch-from-config.ts: DispatchFromConfigParams",
   "src/auto-reply/reply/dispatch-from-config.ts: getDispatcherFinalOutcomeCounts",
   "src/auto-reply/reply/dispatch-from-config.ts: testing",
   "src/auto-reply/reply/dispatch-from-config.types.ts: DispatchReplyFromConfig",
   "src/auto-reply/reply/dispatcher-registry.ts: clearAllDispatchers",
   "src/auto-reply/reply/effective-reply-route.ts: EffectiveReplyRouteContext",
   "src/auto-reply/reply/effective-reply-route.ts: EffectiveReplyRouteEntry",
-  "src/auto-reply/reply/exec.ts: extractExecDirective",
   "src/auto-reply/reply/get-reply-directives-apply.ts: formatModelOverrideResetEvent",
   "src/auto-reply/reply/get-reply-fast-path.ts: markCompleteReplyConfig",
   "src/auto-reply/reply/get-reply-fast-path.ts: withFastReplyConfig",
@@ -1761,97 +1723,40 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/auto-reply/reply/get-reply-run.ts: resolvePromptSessionContextForSystemEvent",
   "src/auto-reply/reply/get-reply-run.ts: resolvePromptSilentReplyConversationType",
   "src/auto-reply/reply/history.ts: appendHistoryEntry",
-  "src/auto-reply/reply/history.ts: HistoryMediaEntry",
   "src/auto-reply/reply/inbound-dedupe.ts: buildInboundDedupeKey",
   "src/auto-reply/reply/inbound-dedupe.ts: InboundDedupeClaimResult",
-  "src/auto-reply/reply/inbound-media.ts: InboundMediaContext",
-  "src/auto-reply/reply/mentions.ts: ExplicitMentionSignal",
   "src/auto-reply/reply/mentions.types.ts: BuildMentionRegexes",
   "src/auto-reply/reply/mentions.types.ts: MatchesMentionPatterns",
   "src/auto-reply/reply/mentions.types.ts: MatchesMentionWithExplicit",
   "src/auto-reply/reply/progress-narrator.ts: createProgressNarrator",
   "src/auto-reply/reply/progress-narrator.ts: ProgressNarrationInput",
   "src/auto-reply/reply/prompt-prelude.ts: buildReplyPromptBodies",
-  "src/auto-reply/reply/prompt-prelude.ts: ReplyPromptEnvelope",
-  "src/auto-reply/reply/prompt-prelude.ts: ReplyPromptEnvelopeBase",
   "src/auto-reply/reply/provider-request-error-classifier.ts: PROVIDER_AUTHENTICATION_ERROR_USER_MESSAGE",
   "src/auto-reply/reply/provider-request-error-classifier.ts: PROVIDER_INTERNAL_ERROR_USER_MESSAGE",
   "src/auto-reply/reply/provider-request-error-classifier.ts: PROVIDER_MODEL_UNAVAILABLE_USER_MESSAGE",
   "src/auto-reply/reply/provider-request-error-classifier.ts: PROVIDER_RATE_LIMIT_OR_QUOTA_ERROR_USER_MESSAGE",
-  "src/auto-reply/reply/provider-request-error-classifier.ts: ProviderRequestErrorClassification",
-  "src/auto-reply/reply/provider-request-error-classifier.ts: ProviderRequestErrorCode",
-  "src/auto-reply/reply/queue.ts: clearFollowupQueue",
-  "src/auto-reply/reply/queue.ts: EnqueueFollowupRunOptions",
-  "src/auto-reply/reply/queue.ts: extractQueueDirective",
-  "src/auto-reply/reply/queue.ts: isFollowupRunDeferredError",
-  "src/auto-reply/reply/queue.ts: QueueDedupeMode",
-  "src/auto-reply/reply/queue.ts: QueueDropPolicy",
-  "src/auto-reply/reply/queue.ts: QueueInsertPosition",
-  "src/auto-reply/reply/queue.ts: QueueMode",
   "src/auto-reply/reply/queue.ts: resetRecentQueuedMessageIdDedupe",
   "src/auto-reply/reply/queue/cleanup.ts: testing",
   "src/auto-reply/reply/queue/drain.ts: resolveFollowupAuthorizationKey",
   "src/auto-reply/reply/queue/enqueue.ts: resetRecentQueuedMessageIdDedupe",
-  "src/auto-reply/reply/queue/state.ts: FollowupQueueState",
-  "src/auto-reply/reply/queue/types.ts: QueueInsertPosition",
   "src/auto-reply/reply/reply-dispatcher.runtime-types.ts: CreateReplyDispatcherWithTyping",
-  "src/auto-reply/reply/reply-dispatcher.ts: ReplyDispatcher",
-  "src/auto-reply/reply/reply-dispatcher.ts: ReplyDispatchKind",
-  "src/auto-reply/reply/reply-operation-run-state.ts: ReplyOperationAdmissionSnapshot",
   "src/auto-reply/reply/reply-operation-run-state.ts: ReplyOptionsWithOperationRunState",
   "src/auto-reply/reply/reply-payloads-dedupe.ts: getMatchingMessagingToolReplyTargets",
-  "src/auto-reply/reply/reply-payloads.ts: shouldDedupeMessagingToolRepliesForRoute",
   "src/auto-reply/reply/reply-run-registry.ts: __testing",
-  "src/auto-reply/reply/reply-run-registry.ts: ReplyBackendCancelReason",
-  "src/auto-reply/reply/reply-run-registry.ts: ReplyBackendKind",
-  "src/auto-reply/reply/reply-run-registry.ts: ReplyBackendQueueMessageMismatch",
-  "src/auto-reply/reply/reply-run-registry.ts: ReplyOperationAbortCode",
-  "src/auto-reply/reply/reply-run-registry.ts: ReplyOperationFailureCode",
-  "src/auto-reply/reply/reply-run-registry.ts: ReplyOperationResult",
-  "src/auto-reply/reply/reply-run-registry.ts: ReplyOperationStaleReason",
-  "src/auto-reply/reply/reply-run-registry.ts: ReplyRunKey",
-  "src/auto-reply/reply/reply-run-registry.ts: ReplyRunRegistry",
   "src/auto-reply/reply/reply-run-registry.ts: testing",
   "src/auto-reply/reply/reply-threading.ts: createReplyToModeFilter",
   "src/auto-reply/reply/reply-threading.ts: resolveConfiguredReplyToMode",
   "src/auto-reply/reply/reply-threading.ts: resolveReplyToModeWithThreading",
-  "src/auto-reply/reply/reply-turn-admission.ts: ReplyTurnAdmission",
-  "src/auto-reply/reply/reply-turn-admission.ts: ReplyTurnKind",
   "src/auto-reply/reply/reply-usage-state.ts: clearReplyUsageStateForTest",
-  "src/auto-reply/reply/session-delivery.ts: LegacyMainDeliveryRetirement",
-  "src/auto-reply/reply/session-fork.ts: ForkedParentSessionEntry",
   "src/auto-reply/reply/session-fork.ts: ForkSessionEntryFromParentParams",
   "src/auto-reply/reply/session-fork.ts: ForkSessionEntryFromParentResult",
   "src/auto-reply/reply/session-fork.ts: ParentForkDecision",
   "src/auto-reply/reply/session-reset-prompt.ts: buildBareSessionResetPrompt",
-  "src/auto-reply/reply/session-updates.ts: drainFormattedSystemEvents",
-  "src/auto-reply/reply/session-updates.ts: resetResolvedSkillsCacheForTests",
-  "src/auto-reply/reply/session.ts: InitSessionStateParams",
-  "src/auto-reply/reply/session.ts: ReplySessionPreprocessingState",
   "src/auto-reply/reply/stage-sandbox-media.ts: appendScpStderrTail",
   "src/auto-reply/reply/stage-sandbox-media.ts: SCP_STDERR_TAIL_CHARS",
   "src/auto-reply/reply/stage-sandbox-media.ts: testing",
   "src/auto-reply/reply/stranded-reply-recovery.ts: STRANDED_REPLY_RETRY_MARKER",
-  "src/auto-reply/reply/typing-mode.ts: TypingModeContext",
-  "src/auto-reply/status.ts: buildStatusMessage",
-  "src/auto-reply/status.ts: CommandsMessageOptions",
-  "src/auto-reply/status.ts: CommandsMessageResult",
-  "src/auto-reply/status.ts: StatusArgs",
-  "src/auto-reply/thinking.shared.ts: ALL_THINKING_LEVELS",
-  "src/auto-reply/thinking.shared.ts: resolveMessagesResponseUsageDefault",
-  "src/auto-reply/thinking.shared.ts: ResponseUsageDefaultConfig",
-  "src/auto-reply/thinking.shared.ts: ResponseUsageInput",
-  "src/auto-reply/thinking.ts: resolveMessagesResponseUsageDefault",
-  "src/auto-reply/thinking.ts: ResponseUsageDefaultConfig",
-  "src/auto-reply/thinking.ts: ResponseUsageInput",
-  "src/auto-reply/thinking.ts: ThinkingLevelOption",
-  "src/auto-reply/types.ts: copyReplyPayloadMetadata",
-  "src/auto-reply/types.ts: markCommandReplyForDelivery",
-  "src/auto-reply/types.ts: markReplyPayloadForSourceSuppressionDelivery",
-  "src/auto-reply/types.ts: PartialReplyPayload",
-  "src/auto-reply/types.ts: setReplyPayloadMetadata",
   "src/auto-reply/usage-bar/template.ts: clearUsageBarTemplateCacheForTest",
-  "src/auto-reply/usage-bar/template.ts: UsageTemplateConfig",
   "src/channels/ids.ts: CHAT_CHANNEL_ALIASES",
   "src/channels/message-access/index.ts: resolveChannelIngressEffectiveAllowFromLists",
   "src/channels/message-access/runtime-types.ts: ChannelIngressSubjectIdentifier",
@@ -2100,77 +2005,41 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/channels/turn/types.ts: InboundEventKind",
   "src/channels/turn/types.ts: ProjectedAllowlistAccessFacts",
   "src/channels/turn/types.ts: ProjectedEventAccessFacts",
-  "src/cli/argv.ts: hasHelpOrVersion",
-  "src/cli/argv.ts: NormalizeRootLogLevelArgvOptions",
-  "src/cli/argv.ts: NormalizeRootNoColorArgvOptions",
   "src/cli/channel-options.ts: testing",
-  "src/cli/clawhub-risk-acknowledgement.ts: ClawHubRiskAcknowledgementCliOptions",
   "src/cli/command-path-policy.ts: resolveCliCatalogCommandPath",
   "src/cli/command-secret-gateway.ts: testing",
   "src/cli/cron-cli/register.cron-simple.ts: loadCronJobForShow",
-  "src/cli/cron-cli/schedule-options.ts: CronEditScheduleRequest",
   "src/cli/daemon-cli/launchd-recovery.ts: LAUNCH_AGENT_RECOVERY_MESSAGE",
   "src/cli/daemon-cli/response.ts: buildDaemonHintItems",
-  "src/cli/daemon-cli/response.ts: DaemonAction",
-  "src/cli/daemon-cli/response.ts: DaemonHintItem",
-  "src/cli/daemon-cli/response.ts: DaemonHintKind",
   "src/cli/gateway-cli/call.ts: gatewayCallOpts",
   "src/cli/gateway-cli/qa-parent-watchdog.ts: QA_PARENT_PID_ENV",
   "src/cli/gateway-cli/qa-parent-watchdog.ts: QA_STAGED_RUNTIME_ROOT_ENV",
   "src/cli/gateway-cli/qa-parent-watchdog.ts: QA_TEMP_ROOT_ENV",
-  "src/cli/gateway-cli/qa-parent-watchdog.ts: QaParentWatchdogHandle",
-  "src/cli/gateway-cli/run.ts: __testing",
   "src/cli/gateway-cli/run.ts: testing",
-  "src/cli/help-format.ts: HelpExample",
-  "src/cli/nodes-camera.ts: CameraClipPayload",
-  "src/cli/nodes-camera.ts: CameraSnapPayload",
   "src/cli/nodes-camera.ts: writeUrlToFile",
-  "src/cli/nodes-screen.ts: ScreenRecordPayload",
-  "src/cli/nodes-screen.ts: ScreenSnapshotPayload",
   "src/cli/npm-resolution.ts: logPinnedNpmSpecMessages",
   "src/cli/npm-resolution.ts: resolvePinnedNpmInstallRecord",
   "src/cli/npm-resolution.ts: resolvePinnedNpmSpec",
-  "src/cli/plugins-cli.ts: PluginUninstallOptions",
   "src/cli/plugins-install-command.ts: loadConfigForInstall",
   "src/cli/plugins-list-format.ts: formatPluginLine",
-  "src/cli/plugins-location-bridges.ts: PersistedBundledPluginRecoveryLocation",
-  "src/cli/plugins-search-command.ts: PluginsSearchOptions",
   "src/cli/plugins-uninstall-command.ts: PluginUninstallOptions",
-  "src/cli/ports.ts: ForceFreePortResult",
   "src/cli/ports.ts: listPortListeners",
   "src/cli/ports.ts: parseLsofOutput",
   "src/cli/ports.ts: PortProcess",
   "src/cli/ports.ts: probePortFree",
-  "src/cli/precomputed-help.ts: PrecomputedSubcommandHelpName",
   "src/cli/precomputed-help.ts: resolvePrecomputedSubcommandHelpCommand",
-  "src/cli/program/command-descriptor-utils.ts: CommandDescriptorCatalog",
-  "src/cli/program/command-descriptor-utils.ts: CommandDescriptorLike",
   "src/cli/program/command-descriptor-utils.ts: getCommandDescriptorNames",
   "src/cli/program/command-descriptor-utils.ts: getCommandsWithSubcommands",
   "src/cli/program/command-descriptor-utils.ts: getParentDefaultHelpCommands",
   "src/cli/program/command-group-descriptors.ts: resolveCommandGroupEntries",
-  "src/cli/program/command-group-descriptors.ts: ResolvedCommandGroupEntry",
-  "src/cli/program/command-registry.ts: CommandRegistration",
-  "src/cli/program/command-registry.ts: getCoreCliCommandDescriptors",
   "src/cli/program/command-registry.ts: getCoreCliCommandNames",
   "src/cli/program/command-registry.ts: getCoreCliCommandsWithSubcommands",
   "src/cli/program/command-registry.ts: registerCoreCliCommands",
   "src/cli/program/command-tree.ts: removeCommand",
-  "src/cli/program/core-command-descriptors.ts: CoreCliCommandDescriptor",
-  "src/cli/program/register-command-groups.ts: CommandGroupPlaceholderOption",
-  "src/cli/program/register.subclis-core.ts: getSubCliCommandsWithSubcommands",
-  "src/cli/program/register.subclis.ts: getSubCliCommandsWithSubcommands",
   "src/cli/program/register.subclis.ts: getSubCliEntries",
-  "src/cli/program/routes.ts: RouteSpec",
-  "src/cli/shared/parse-port.ts: MAX_TCP_PORT",
-  "src/cli/shared/parse-port.ts: parseTcpPort",
   "src/cli/startup-metadata.ts: testing",
-  "src/cli/startup-trace.ts: GatewayStartupTraceSource",
   "src/cli/tagline.ts: DEFAULT_TAGLINE",
-  "src/cli/update-cli/plugin-payload-validation.ts: PluginPayloadSmokeFailureReason",
-  "src/cli/update-cli/plugin-payload-validation.ts: PluginPayloadSmokeResult",
   "src/cli/update-cli/progress.ts: inferUpdateFailureHints",
-  "src/cli/update-cli/progress.ts: ProgressController",
   "src/cli/update-cli/update-command.ts: buildInvalidConfigPostCoreUpdateResult",
   "src/cli/update-cli/update-command.ts: collectMissingPluginInstallPayloads",
   "src/cli/update-cli/update-command.ts: formatPostUpdateGatewayRecoveryInstructions",
@@ -2187,28 +2056,16 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/commands/auth-choice-options.ts: buildAuthChoiceOptions",
   "src/commands/auth-choice.model-check.ts: DefaultModelAuthStatus",
   "src/commands/auth-choice.model-check.ts: DefaultModelCatalogFacts",
-  "src/commands/auth-choice.model-check.ts: ModelRouteObservation",
-  "src/commands/auth-token.ts: ANTHROPIC_SETUP_TOKEN_PREFIX",
-  "src/commands/backup-shared.ts: BackupAssetKind",
-  "src/commands/backup-shared.ts: BackupPlan",
-  "src/commands/backup-shared.ts: BackupSkipReason",
   "src/commands/backup-shared.ts: encodeAbsolutePathForBackupArchive",
   "src/commands/backup-shared.ts: formatBackupArchiveTimestamp",
   "src/commands/backup-shared.ts: resolveBackupPlanFromPaths",
-  "src/commands/backup-shared.ts: SkippedBackupAsset",
-  "src/commands/configure.ts: buildGatewayAuthConfig",
-  "src/commands/configure.ts: CONFIGURE_WIZARD_SECTIONS",
-  "src/commands/crestodian-with-inference.ts: CrestodianWithInferenceDeps",
   "src/commands/daemon-install-helpers.ts: collectPreservedExistingServiceEnvVars",
-  "src/commands/daemon-install-helpers.ts: resolveGatewayDevMode",
   "src/commands/daemon-install-plan.shared.ts: resolveDaemonOpenClawBinDir",
   "src/commands/daemon-install-plan.shared.ts: resolveGatewayDevMode",
   "src/commands/doctor-auth-oauth-sidecar.ts: testing",
   "src/commands/doctor-auth.ts: formatOAuthRefreshFailureDoctorLine",
   "src/commands/doctor-auth.ts: legacyCodexProviderOverrideToHealthFinding",
   "src/commands/doctor-auth.ts: resolveUnusableProfileHint",
-  "src/commands/doctor-completion.ts: DoctorCompletionOptions",
-  "src/commands/doctor-completion.ts: ShellCompletionStatusOptions",
   "src/commands/doctor-config-analysis.ts: collectImplicitFallbackClobberWarnings",
   "src/commands/doctor-disk-space.ts: buildDiskSpaceWarnings",
   "src/commands/doctor-heartbeat-main-session-repair.ts: clearTuiLastSessionPointers",
@@ -2219,38 +2076,17 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/commands/doctor-platform-notes.ts: collectMacLaunchAgentOverrideWarning",
   "src/commands/doctor-platform-notes.ts: collectMacLaunchctlGatewayEnvOverrideWarning",
   "src/commands/doctor-platform-notes.ts: collectMacStaleOpenClawUpdateLaunchdJobsWarning",
-  "src/commands/doctor-plugin-registry.ts: PluginRegistryHealthIssue",
   "src/commands/doctor-sandbox.ts: resolveSandboxScript",
   "src/commands/doctor-session-snapshots.ts: resolveSessionSnapshotBundledSkillsDir",
   "src/commands/doctor-session-snapshots.ts: scanSessionStoreForStaleRuntimeSnapshotPaths",
-  "src/commands/doctor-session-snapshots.ts: SessionSnapshotHealthIssue",
-  "src/commands/doctor-session-sqlite-github-issue.ts: SessionSqliteGithubIssueCreateResult",
-  "src/commands/doctor-session-sqlite-migration-run.ts: listSessionSqliteMigrationManifestPaths",
-  "src/commands/doctor-session-sqlite-migration-run.ts: readSessionSqliteMigrationManifest",
   "src/commands/doctor-session-sqlite-migration-run.ts: SessionSqliteMigrationManifest",
-  "src/commands/doctor-session-sqlite-migration-run.ts: sessionSqliteMigrationTargetKey",
-  "src/commands/doctor-session-sqlite-migration-run.ts: SessionSqliteMigrationTargetManifest",
-  "src/commands/doctor-session-sqlite-readers.ts: ReadOnlySqliteDbStatsResult",
-  "src/commands/doctor-session-sqlite-readers.ts: ReadOnlySqliteExactSessionEntryResult",
-  "src/commands/doctor-session-sqlite-readers.ts: ReadOnlySqliteSessionEntriesResult",
-  "src/commands/doctor-session-sqlite-readers.ts: ReadOnlySqliteSessionSummary",
-  "src/commands/doctor-session-sqlite-readers.ts: ReadOnlySqliteTranscriptEventCountResult",
-  "src/commands/doctor-session-sqlite-readers.ts: TranscriptEventCountResult",
-  "src/commands/doctor-session-sqlite-recover-report.ts: SessionSqliteRecoverTargetValidator",
-  "src/commands/doctor-session-sqlite-types.ts: DoctorSessionSqliteCorruptRecovery",
-  "src/commands/doctor-session-sqlite-types.ts: DoctorSessionSqliteDbStats",
-  "src/commands/doctor-session-sqlite-types.ts: DoctorSessionSqliteLargestSession",
   "src/commands/doctor-session-state-providers.ts: applySessionRouteStateRepair",
   "src/commands/doctor-session-state-providers.ts: resolveConfiguredDoctorSessionStateRoute",
   "src/commands/doctor-session-state-providers.ts: scanSessionRouteStateOwners",
   "src/commands/doctor-session-state-providers.ts: storeMayContainPluginSessionRouteState",
   "src/commands/doctor-session-transcripts.ts: repairBrokenSessionTranscriptFile",
-  "src/commands/doctor-session-transcripts.ts: SessionTranscriptHealthIssue",
-  "src/commands/doctor-skills.ts: collectUnavailableAgentSkills",
   "src/commands/doctor-skills.ts: describeGhConfigDirHintFromDiscovery",
-  "src/commands/doctor-skills.ts: disableUnavailableSkillsInConfig",
   "src/commands/doctor-skills.ts: formatUnavailableSkillDoctorLines",
-  "src/commands/doctor-sqlite-compact.ts: DoctorSqliteCompactResult",
   "src/commands/doctor-state-sqlite-compact.ts: DoctorStateSqliteCompactReport",
   "src/commands/doctor-ui.ts: UiProtocolFreshnessIssue",
   "src/commands/doctor-whatsapp-responsiveness.ts: listLocalTuiProcesses",
@@ -2259,104 +2095,47 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/commands/doctor/cron/legacy-delivery.ts: buildDeliveryPatchFromLegacyPayload",
   "src/commands/doctor/cron/legacy-delivery.ts: hasLegacyDeliveryHints",
   "src/commands/doctor/cron/legacy-delivery.ts: mergeLegacyDeliveryInto",
-  "src/commands/doctor/cron/legacy-store-migration.ts: LegacyCronArchiveResult",
-  "src/commands/doctor/cron/payload-migration.ts: UnresolvedAgentTurnShellToolPromptKind",
   "src/commands/doctor/cron/warnings.ts: collectCronDeliveryTargetAdvisory",
-  "src/commands/doctor/shared/channel-plugin-blockers.ts: ChannelPluginBlockerHit",
   "src/commands/doctor/shared/codex-native-assets.ts: scanCodexNativeAssets",
-  "src/commands/doctor/shared/codex-route-warnings.ts: DisabledCodexPluginRouteIssue",
   "src/commands/doctor/shared/codex-route-warnings.ts: repairCodexSessionStoreRoutes",
   "src/commands/doctor/shared/context-engine-host-compat.ts: collectConfiguredContextEngineAgentRunHosts",
-  "src/commands/doctor/shared/exec-safe-bins.ts: ExecSafeBinCoverageHit",
-  "src/commands/doctor/shared/exec-safe-bins.ts: ExecSafeBinTrustedDirHintHit",
-  "src/commands/doctor/shared/legacy-config-state-migration-input.ts: StateMigrationConfigInput",
   "src/commands/doctor/shared/legacy-oauth-sidecar.ts: legacyOAuthSidecarInternalTestUtils",
-  "src/commands/doctor/shared/legacy-tools-by-sender.ts: LegacyToolsBySenderKeyHit",
-  "src/commands/doctor/shared/legacy-web-search-migrate.ts: resolveLegacyGrokWebSearchModelTarget",
-  "src/commands/doctor/shared/missing-configured-plugin-install.ts: ConfiguredPluginInstallHealthIssue",
-  "src/commands/doctor/shared/missing-configured-plugin-install.ts: RepairMissingPluginInstallsResult",
-  "src/commands/doctor/shared/plugin-dependency-cleanup.ts: LegacyPluginDependencyStateIssue",
   "src/commands/doctor/shared/plugin-dependency-cleanup.ts: testing",
   "src/commands/doctor/shared/plugin-registry-migration.ts: FORCE_PLUGIN_REGISTRY_MIGRATION_ENV",
-  "src/commands/doctor/shared/plugin-registry-migration.ts: PluginRegistryInstallMigrationPreflight",
-  "src/commands/doctor/shared/plugin-registry-migration.ts: PluginRegistryInstallMigrationPreflightAction",
-  "src/commands/doctor/shared/plugin-registry-migration.ts: PluginRegistryInstallMigrationResult",
   "src/commands/doctor/shared/plugin-runtime-symlinks.ts: collectStalePluginRuntimeSymlinks",
-  "src/commands/doctor/shared/plugin-runtime-symlinks.ts: PluginRuntimeSymlinkOptions",
-  "src/commands/doctor/shared/plugin-runtime-symlinks.ts: StalePluginRuntimeSymlink",
   "src/commands/doctor/shared/plugin-runtime-symlinks.ts: stalePluginRuntimeSymlinkToHealthFinding",
   "src/commands/doctor/shared/preview-warnings.ts: collectChannelBoundMessageToolPolicyWarnings",
   "src/commands/doctor/shared/preview-warnings.ts: collectProfileConfiguredToolSectionWarnings",
   "src/commands/doctor/shared/preview-warnings.ts: collectVisibleReplyToolPolicyWarnings",
-  "src/commands/doctor/shared/preview-warnings.ts: DoctorPreviewNotes",
   "src/commands/doctor/shared/release-configured-plugin-installs.ts: collectReleaseConfiguredPluginIds",
   "src/commands/doctor/shared/release-configured-plugin-installs.ts: shouldRunConfiguredPluginInstallReleaseStep",
   "src/commands/doctor/shared/stale-auth-order.ts: repairStaleConfiguredAuthOrders",
-  "src/commands/doctor/shared/stale-auth-order.ts: scanStaleConfiguredAuthOrders",
   "src/commands/doctor/shared/stale-oauth-profile-shadows.ts: testing",
-  "src/commands/doctor/shared/stale-subagent-allowlist.ts: StaleSubagentAllowlistHit",
-  "src/commands/doctor/types.ts: DoctorAllowFromEntry",
-  "src/commands/gateway-status/output.ts: GatewayStatusWarning",
-  "src/commands/health.types.ts: ConfigReloadHealthSummary",
   "src/commands/migrate/selection.ts: MIGRATION_PLUGIN_NOT_SELECTED_REASON",
   "src/commands/migrate/selection.ts: MIGRATION_SKILL_NOT_SELECTED_REASON",
   "src/commands/migrate/selection.ts: MIGRATION_SKILL_SELECTION_ACCEPT",
   "src/commands/migrate/selection.ts: MIGRATION_SKILL_SELECTION_TOGGLE_ALL_OFF",
   "src/commands/migrate/selection.ts: MIGRATION_SKILL_SELECTION_TOGGLE_ALL_ON",
-  "src/commands/models/list.auth-index.ts: CreateModelListAuthIndexParams",
-  "src/commands/models/load-config.ts: LoadedModelsConfig",
-  "src/commands/models/shared.ts: PrimaryFallbackConfig",
-  "src/commands/models/shared.ts: UpdateConfigContext",
-  "src/commands/onboard-custom-config.ts: inferCustomModelSupportsImageInput",
-  "src/commands/onboard-custom.ts: applyCustomApiConfig",
-  "src/commands/onboard-custom.ts: buildAnthropicVerificationProbeRequest",
-  "src/commands/onboard-custom.ts: buildOpenAiVerificationProbeRequest",
-  "src/commands/onboard-custom.ts: CustomApiError",
-  "src/commands/onboard-custom.ts: inferCustomModelSupportsImageInput",
-  "src/commands/onboard-custom.ts: parseNonInteractiveCustomApiFlags",
-  "src/commands/onboard-custom.ts: resolveCustomModelImageInputInference",
-  "src/commands/onboard-custom.ts: resolveCustomProviderId",
-  "src/commands/onboard-inference.ts: DetectInferenceBackendsDeps",
-  "src/commands/onboard-inference.ts: DetectInferenceBackendsOptions",
   "src/commands/onboard-inference.ts: detectNativeCodexAppServer",
-  "src/commands/onboard-inference.ts: DetectNativeCodexAppServerOptions",
-  "src/commands/onboard-inference.ts: InferenceBackendCandidate",
-  "src/commands/onboard-non-interactive/api-keys.ts: NonInteractiveApiKeySource",
   "src/commands/onboard-non-interactive/local.ts: resolveGatewayHealthProbeToken",
   "src/commands/onboard-non-interactive/local.ts: resolveInstallDaemonGatewayHealthTiming",
   "src/commands/onboard-remote-gateway.ts: RemoteGatewayInferenceOnboardingDeps",
   "src/commands/onboard-remote-gateway.ts: RemoteGatewayInferenceTarget",
   "src/commands/onboard-skills.ts: testing",
-  "src/commands/onboard.ts: onboardCommand",
   "src/commands/onboarding-plugin-install.ts: testing",
-  "src/commands/promos/claim.ts: PromosClaimOptions",
-  "src/commands/reset.ts: ResetOptions",
-  "src/commands/reset.ts: ResetScope",
-  "src/commands/runtime-plugin-install.ts: RuntimePluginEnsureParams",
-  "src/commands/runtime-plugin-install.ts: RuntimePluginInstallDescriptor",
   "src/commands/runtime-plugin-install.ts: RuntimePluginInstallResult",
-  "src/commands/runtime-plugin-install.ts: RuntimePluginModelSelectionHelpers",
-  "src/commands/runtime-plugin-install.ts: RuntimePluginRepairParams",
-  "src/commands/runtime-plugin-install.ts: RuntimePluginSelection",
   "src/commands/session-store-targets.ts: resolveSessionStoreTargets",
-  "src/commands/sessions-compact.ts: SessionsCompactCliOptions",
   "src/commands/sessions-tail.ts: setSessionsTailFollowIntervalMsForTests",
-  "src/commands/sessions.ts: __testing",
   "src/commands/sessions.ts: testing",
-  "src/commands/status-all/channels.ts: ChannelRow",
-  "src/commands/status-runtime-shared.ts: resolveStatusLastHeartbeat",
-  "src/commands/status-runtime-shared.ts: resolveStatusRuntimeDetails",
   "src/commands/status.command.ts: resolvePairingRecoveryContext",
   "src/commands/status.scan.config-shared.ts: resolveStatusScanColdStart",
   "src/commands/status.scan.config-shared.ts: shouldSkipStatusScanMissingConfigFastPath",
-  "src/commands/uninstall.ts: UninstallOptions",
   "src/commitments/extraction.ts: validateCommitmentCandidates",
   "src/commitments/runtime.ts: configureCommitmentExtractionRuntime",
   "src/commitments/runtime.ts: drainCommitmentExtractionQueue",
   "src/commitments/runtime.ts: resetCommitmentExtractionRuntimeForTests",
   "src/commitments/store.ts: loadCommitmentStore",
   "src/commitments/store.ts: saveCommitmentStore",
-  "src/commitments/types.ts: CommitmentDueWindow",
   "src/compat/legacy-names.ts: MACOS_APP_SOURCES_DIR",
   "src/compat/legacy-names.ts: PROJECT_NAME",
   "src/config/backup-rotation.ts: cleanOrphanBackups",
@@ -2573,119 +2352,58 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/config/zod-schema.proxy.ts: ProxyLoopbackModeSchema",
   "src/config/zod-schema.ts: installZodDefaultLocale",
   "src/config/zod-schema.ts: McpServerSchema",
-  "src/context-engine/host-compat.ts: ContextEngineHostSupportEvaluation",
-  "src/context-engine/quarantine-health.ts: PersistedContextEngineRuntimeQuarantine",
   "src/context-engine/registry.ts: clearContextEngineRuntimeQuarantine",
-  "src/context-engine/registry.ts: ContextEngineRegistration",
-  "src/context-engine/registry.ts: ContextEngineRegistrationLifecycle",
   "src/context-engine/registry.ts: ContextEngineRegistrationResult",
-  "src/context-engine/registry.ts: ContextEngineRuntimeQuarantine",
   "src/context-engine/registry.ts: getContextEngineFactory",
   "src/context-engine/registry.ts: listContextEngineIds",
-  "src/context-engine/types.ts: ContextEngineTranscriptStorageInfo",
-  "src/crestodian/agent-turn.ts: CRESTODIAN_AGENT_ID",
   "src/crestodian/agent-turn.ts: CrestodianAgentTurnDeps",
   "src/crestodian/agent-turn.ts: CrestodianAgentTurnDirective",
-  "src/crestodian/agent-turn.ts: CrestodianAgentTurnReply",
   "src/crestodian/agent-turn.ts: runCrestodianAgentTurnWithDeps",
-  "src/crestodian/assistant-prompts.ts: CRESTODIAN_ASSISTANT_MAX_TOKENS",
-  "src/crestodian/chat-engine.ts: CrestodianChatReply",
-  "src/crestodian/chat-engine.ts: CrestodianChatReplyAction",
-  "src/crestodian/inference-error.ts: CrestodianInferenceStage",
-  "src/crestodian/inference-route.ts: resolveCrestodianConfiguredRoute",
-  "src/crestodian/rescue-message.ts: CrestodianRescueMessageInput",
-  "src/crestodian/verified-inference.ts: CrestodianConfiguredRouteIdentity",
-  "src/crestodian/verified-inference.ts: crestodianRouteIdentity",
   "src/cron/command-output-summary.ts: cronCommandSummaryNeedsExternalRedaction",
   "src/cron/delivery-context.ts: cronDeliveryFromContext",
   "src/cron/delivery-preview.ts: resolveCronDeliveryPreview",
-  "src/cron/delivery.ts: CronAnnounceTarget",
-  "src/cron/delivery.ts: CronDeliveryPlan",
-  "src/cron/delivery.ts: CronFailureDeliveryPlan",
-  "src/cron/delivery.ts: CronFailureDestinationInput",
   "src/cron/isolated-agent.ts: RunCronAgentTurnResult",
-  "src/cron/isolated-agent/delivery-dispatch.ts: DispatchCronDeliveryState",
   "src/cron/isolated-agent/delivery-dispatch.ts: getCompletedDirectCronDeliveriesCountForTests",
   "src/cron/isolated-agent/delivery-dispatch.ts: resetCompletedDirectCronDeliveriesForTests",
-  "src/cron/isolated-agent/delivery-dispatch.ts: SuccessfulDeliveryTarget",
-  "src/cron/isolated-agent/helpers.ts: CronPayloadOutcome",
   "src/cron/isolated-agent/helpers.ts: pickDeliverablePayloads",
   "src/cron/isolated-agent/helpers.ts: pickLastDeliverablePayload",
   "src/cron/isolated-agent/helpers.ts: pickSummaryFromPayloads",
-  "src/cron/isolated-agent/model-selection.ts: ResolveCronModelSelectionParams",
-  "src/cron/isolated-agent/model-selection.ts: ResolveCronModelSelectionResult",
-  "src/cron/isolated-agent/run-execution.runtime.ts: getCliSessionId",
   "src/cron/isolated-agent/run-executor.ts: createCronPromptExecutor",
-  "src/cron/isolated-agent/run-session-state.ts: MutableCronSessionEntry",
-  "src/cron/isolated-agent/run.runtime.ts: DEFAULT_IDENTITY_FILENAME",
-  "src/cron/isolated-agent/run.runtime.ts: ResolvedAgentConfig",
   "src/cron/isolated-agent/run.ts: resolveCronDeliveryContext",
   "src/cron/isolated-agent/run.ts: RunCronAgentTurnResult",
-  "src/cron/isolated-agent/source-delivery-fallback.ts: CronSourceDeliveryResolvedTarget",
   "src/cron/job-session-bindings.ts: CronJobSessionBinding",
-  "src/cron/persisted-shape.ts: InvalidPersistedCronJobReason",
-  "src/cron/retry-hint.ts: CronRetryHint",
   "src/cron/run-diagnostics.ts: MISSING_WEB_SEARCH_PROVIDER_DIAGNOSTIC_MESSAGE",
   "src/cron/run-log.ts: getPendingCronRunLogWriteCountForTests",
   "src/cron/schedule.ts: clearCronScheduleCacheForTest",
   "src/cron/schedule.ts: getCronScheduleCacheMaxForTest",
   "src/cron/schedule.ts: getCronScheduleCacheSizeForTest",
   "src/cron/schedule.ts: hasCronInCacheForTest",
-  "src/cron/service.ts: CronServiceDeps",
-  "src/cron/service/agent-watchdog.ts: CronAgentWatchdog",
-  "src/cron/service/state.ts: CronDeclarativeAddResult",
-  "src/cron/service/state.ts: CronServiceDepsInternal",
   "src/cron/service/timeout-policy.ts: AGENT_TURN_SAFETY_TIMEOUT_MS",
   "src/cron/service/timeout-policy.ts: DEFAULT_JOB_TIMEOUT_MS",
-  "src/cron/service/timer.ts: CronTriggerEvalOutcome",
-  "src/cron/service/timer.ts: DEFAULT_JOB_TIMEOUT_MS",
   "src/cron/service/timer.ts: executeJobCore",
-  "src/cron/service/timer.ts: failureNotificationDeliveryFromJobState",
-  "src/cron/service/timer.ts: normalizeCronRunErrorText",
   "src/cron/service/timer.ts: onTimer",
-  "src/cron/service/timer.ts: wake",
   "src/cron/session-reaper.ts: resetReaperThrottle",
   "src/cron/session-reaper.ts: resolveRetentionMs",
   "src/cron/stagger.ts: isRecurringTopOfHourCronExpr",
-  "src/cron/store.ts: CronQuarantineFile",
   "src/cron/store/row-codec.ts: bindScheduleColumns",
   "src/cron/store/row-codec.ts: scheduleFromRow",
-  "src/cron/types.ts: CronCompletionDestination",
-  "src/cron/types.ts: CronFailureDestination",
-  "src/cron/types.ts: CronFailureDestinationPatch",
-  "src/cron/types.ts: CronSessionTarget",
-  "src/cron/types.ts: CronUsageSummary",
-  "src/cron/types.ts: CronWakeMode",
   "src/daemon/constants.ts: formatGatewayServiceDescription",
   "src/daemon/constants.ts: GATEWAY_SYSTEMD_SERVICE_NAME",
   "src/daemon/constants.ts: GATEWAY_WINDOWS_TASK_NAME",
   "src/daemon/constants.ts: normalizeGatewayProfile",
   "src/daemon/gateway-entrypoint.ts: buildGatewayInstallEntrypointCandidates",
-  "src/daemon/launchd-current-service.ts: CurrentProcessLaunchdServiceLabelOptions",
   "src/daemon/launchd-plist.ts: LAUNCH_AGENT_PROCESS_TYPE",
   "src/daemon/launchd-plist.ts: LAUNCH_AGENT_STDIN_PATH",
   "src/daemon/launchd-plist.ts: LAUNCH_AGENT_THROTTLE_INTERVAL_SECONDS",
   "src/daemon/launchd-plist.ts: LAUNCH_AGENT_UMASK_DECIMAL",
-  "src/daemon/launchd-restart-handoff.ts: isCurrentProcessLaunchdServiceLabel",
   "src/daemon/restart-logs.ts: GATEWAY_RESTART_LOG_FILENAME",
-  "src/daemon/restart-logs.ts: GatewayLogPaths",
   "src/daemon/schtasks.ts: parseSchtasksQuery",
-  "src/daemon/schtasks.ts: ScheduledTaskInfo",
   "src/daemon/service-env.ts: buildMinimalServicePath",
   "src/daemon/service-env.ts: getMinimalServicePathParts",
   "src/daemon/service-env.ts: isNodeVersionManagerRuntime",
   "src/daemon/service-env.ts: resolveLinuxSystemCaBundle",
   "src/daemon/service-layout.ts: resolveGatewayServiceEntrypoint",
-  "src/daemon/service-runtime.ts: GatewayServiceSystemdRuntime",
   "src/daemon/service.ts: GatewayServiceControlArgs",
-  "src/daemon/service.ts: GatewayServiceEnv",
-  "src/daemon/service.ts: GatewayServiceEnvArgs",
-  "src/daemon/service.ts: GatewayServiceManageArgs",
-  "src/daemon/service.ts: GatewayServiceReadOptions",
-  "src/daemon/service.ts: GatewayServiceRestartResult",
-  "src/daemon/service.ts: GatewayServiceStageArgs",
-  "src/daemon/service.ts: GatewayServiceStartResult",
-  "src/daemon/systemd.ts: InstalledSystemdGatewayScope",
   "src/daemon/systemd.ts: parseSystemdShow",
   "src/entry.compile-cache.ts: buildOpenClawCompileCacheRespawnPlan",
   "src/entry.compile-cache.ts: isNodeVersionAffectedByCompileCacheDeadlock",
@@ -2693,52 +2411,18 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/entry.compile-cache.ts: resolveOpenClawCompileCacheDirectory",
   "src/entry.compile-cache.ts: runOpenClawCompileCacheRespawnPlan",
   "src/entry.compile-cache.ts: shouldEnableOpenClawCompileCache",
-  "src/fleet/backup.runtime.ts: DEFAULT_FLEET_BACKUP_MAX_BYTES",
-  "src/fleet/backup.runtime.ts: FLEET_BACKUP_MAX_ENTRIES",
-  "src/fleet/backup.runtime.ts: FleetBackupResult",
-  "src/fleet/backup.runtime.ts: FleetRestoreResult",
   "src/fleet/backup.runtime.ts: resolveRestoreOwner",
-  "src/fleet/cell-profile.ts: CellContainerArgOptions",
   "src/fleet/cell-profile.ts: FLEET_BASE_PORT",
   "src/fleet/cell-profile.ts: FLEET_CONTAINER_AUTH_SECRET_DIR",
-  "src/fleet/cell-profile.ts: FLEET_CONTAINER_HOME",
   "src/fleet/cell-profile.ts: FLEET_CONTAINER_STATE_DIR",
-  "src/fleet/cell-profile.ts: FLEET_MANAGED_ENV_KEYS",
-  "src/fleet/containers.runtime.ts: __test",
-  "src/fleet/containers.runtime.ts: CellContainerProfile",
   "src/fleet/containers.runtime.ts: FleetContainerCommandExecutor",
-  "src/fleet/containers.runtime.ts: FleetContainerCommandOptions",
-  "src/fleet/containers.runtime.ts: FleetContainerCommandResult",
-  "src/fleet/containers.runtime.ts: FleetContainerLogsOptions",
-  "src/fleet/containers.runtime.ts: FleetContainerRuntimeName",
   "src/fleet/containers.runtime.ts: FleetContainerStreamExecutor",
-  "src/fleet/containers.runtime.ts: FleetContainerStreamResult",
-  "src/fleet/doctor.runtime.ts: FleetDoctorCellReport",
-  "src/fleet/doctor.runtime.ts: FleetDoctorFinding",
-  "src/fleet/registry.ts: FleetCellOperationLease",
   "src/fleet/registry.ts: ReserveFleetCellParams",
-  "src/fleet/service-support.runtime.ts: FleetHealthResult",
-  "src/fleet/service-support.runtime.ts: rebuildInspectedEnvironment",
-  "src/fleet/service-support.runtime.ts: requirePidsLimit",
-  "src/fleet/service-support.runtime.ts: requirePositiveResource",
-  "src/fleet/service.runtime.ts: FleetActionResult",
-  "src/fleet/service.runtime.ts: FleetCreateResult",
-  "src/fleet/service.runtime.ts: FleetListEntry",
-  "src/fleet/service.runtime.ts: FleetService",
   "src/fleet/service.runtime.ts: FleetServiceOptions",
-  "src/fleet/service.runtime.ts: FleetStatusResult",
   "src/flows/doctor-health-contributions.ts: createDoctorHealthContribution",
   "src/flows/doctor-health-contributions.ts: resolveDoctorHealthContributions",
   "src/flows/doctor-health-contributions.ts: shouldSkipLegacyUpdateDoctorConfigWrite",
-  "src/flows/doctor-lint-flow.ts: DoctorLintRunResult",
-  "src/flows/doctor-repair-flow.ts: DoctorRepairRunOptions",
-  "src/flows/doctor-repair-flow.ts: DoctorRepairRunResult",
-  "src/flows/doctor-tool-result-cap-advice.ts: ToolResultCapDoctorAdviceParams",
-  "src/flows/doctor-tool-result-cap-advice.ts: ToolResultCapDoctorIssue",
   "src/flows/health-check-adapter.ts: defineSplitHealthCheck",
-  "src/flows/health-check-runner-types.ts: HealthCheckRunContext",
-  "src/flows/health-check-runner-types.ts: HealthCheckSelectionOptions",
-  "src/flows/health-checks.ts: HealthCheckMode",
   "src/gateway/active-sessions-shutdown-tracker.ts: ActiveSessionForShutdown",
   "src/gateway/active-sessions-shutdown-tracker.ts: clearActiveSessionsForShutdownTracker",
   "src/gateway/approval-session-audience.ts: ApprovalSessionAudienceSources",
@@ -2933,6 +2617,9 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/gateway/server-methods/agent-timestamp.ts: TimestampInjectionOptions",
   "src/gateway/server-methods/agents-config-mutations.ts: AgentDeleteMutationResult",
   "src/gateway/server-methods/attachment-normalize.ts: RpcAttachmentInput",
+  "src/gateway/server-methods/chat-origin-routing.ts: ChatSendDeliveryEntry",
+  "src/gateway/server-methods/chat-origin-routing.ts: ChatSendOriginatingRoute",
+  "src/gateway/server-methods/chat-server-timing.ts: ChatSendAckServerTiming",
   "src/gateway/server-methods/chat-transcript-inject.ts: GatewayInjectedAbortMeta",
   "src/gateway/server-methods/chat-transcript-inject.ts: GatewayInjectedTranscriptAppendResult",
   "src/gateway/server-methods/config-write-flow.ts: ConfigWriteOptions",
@@ -3059,21 +2746,9 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/hooks/bundled/session-memory/transcript.ts: sanitizeSessionMemoryTranscriptText",
   "src/hooks/gmail-setup-utils.ts: resetGmailSetupUtilsCachesForTest",
   "src/hooks/gmail-setup-utils.ts: resolvePythonExecutablePath",
-  "src/hooks/gmail-watcher-lifecycle.ts: GMailWatcherLog",
-  "src/hooks/gmail-watcher.ts: GmailWatcherStartResult",
-  "src/hooks/hooks-status.ts: HookInstallOption",
-  "src/hooks/hooks-status.ts: HookStatusConfigCheck",
-  "src/hooks/hooks.ts: AgentBootstrapHookContext",
-  "src/hooks/hooks.ts: createHookEvent",
-  "src/hooks/install.ts: HookInstallLogger",
   "src/hooks/install.ts: installHooksFromArchive",
   "src/hooks/internal-hook-types.ts: KNOWN_INTERNAL_HOOK_EVENT_KEYS",
   "src/hooks/module-loader.ts: resolveFileModuleUrl",
-  "src/hooks/update.ts: HookPackUpdateIntegrityDriftParams",
-  "src/hooks/update.ts: HookPackUpdateLogger",
-  "src/hooks/update.ts: HookPackUpdateOutcome",
-  "src/hooks/update.ts: HookPackUpdateStatus",
-  "src/hooks/update.ts: HookPackUpdateSummary",
   "src/hooks/workspace.ts: loadHookEntriesFromDir",
   "src/image-generation/runtime-types.ts: ListRuntimeImageGenerationProvidersParams",
   "src/image-generation/runtime-types.ts: RuntimeImageGenerationProvider",
@@ -3459,39 +3134,14 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/infra/windows-shell-command.ts: RebuiltShellCommandResult",
   "src/infra/windows-shell-command.ts: ShellSegmentRenderResult",
   "src/llm/providers/stream-wrappers/anthropic-family-cache-semantics.ts: isAnthropicBedrockModel",
-  "src/llm/providers/stream-wrappers/moonshot.ts: createMoonshotThinkingWrapper",
-  "src/llm/providers/stream-wrappers/moonshot.ts: resolveMoonshotThinkingKeep",
-  "src/llm/providers/stream-wrappers/moonshot.ts: resolveMoonshotThinkingType",
-  "src/llm/providers/stream-wrappers/reasoning-effort-utils.ts: ReasoningEffort",
-  "src/llm/stream.ts: getEnvApiKey",
-  "src/llm/utils/node-http-proxy.ts: NodeHttpProxyAgents",
   "src/llm/utils/node-http-proxy.ts: resolveHttpProxyUrlForTarget",
   "src/llm/utils/node-http-proxy.ts: UNSUPPORTED_PROXY_PROTOCOL_MESSAGE",
-  "src/llm/utils/oauth/anthropic.ts: loginAnthropic",
   "src/llm/utils/oauth/anthropic.ts: refreshAnthropicToken",
   "src/llm/utils/oauth/anthropic.ts: testing",
-  "src/llm/utils/oauth/github-copilot.ts: getGitHubCopilotBaseUrl",
-  "src/llm/utils/oauth/github-copilot.ts: loginGitHubCopilot",
-  "src/llm/utils/oauth/github-copilot.ts: normalizeDomain",
   "src/llm/utils/oauth/github-copilot.ts: refreshGitHubCopilotToken",
   "src/llm/utils/oauth/github-copilot.ts: testing",
-  "src/llm/utils/oauth/index.ts: anthropicOAuthProvider",
-  "src/llm/utils/oauth/index.ts: getGitHubCopilotBaseUrl",
-  "src/llm/utils/oauth/index.ts: githubCopilotOAuthProvider",
-  "src/llm/utils/oauth/index.ts: loginAnthropic",
-  "src/llm/utils/oauth/index.ts: loginGitHubCopilot",
-  "src/llm/utils/oauth/index.ts: loginOpenAICodex",
-  "src/llm/utils/oauth/index.ts: normalizeDomain",
-  "src/llm/utils/oauth/index.ts: openaiCodexOAuthProvider",
-  "src/llm/utils/oauth/index.ts: refreshAnthropicToken",
-  "src/llm/utils/oauth/index.ts: refreshGitHubCopilotToken",
-  "src/llm/utils/oauth/index.ts: refreshOpenAICodexToken",
   "src/llm/utils/oauth/openai-chatgpt.ts: loginOpenAICodex",
   "src/llm/utils/oauth/openai-chatgpt.ts: refreshOpenAICodexToken",
-  "src/llm/utils/oauth/types.ts: OAuthAuthInfo",
-  "src/llm/utils/oauth/types.ts: OAuthProviderInfo",
-  "src/llm/utils/oauth/types.ts: OAuthSelectOption",
-  "src/llm/utils/oauth/types.ts: OAuthSelectPrompt",
   "src/logging/diagnostic-phase.ts: recordDiagnosticPhase",
   "src/logging/diagnostic-run-activity.ts: markDiagnosticModelStartedForTest",
   "src/logging/diagnostic-run-activity.ts: markDiagnosticRunProgressForTest",
@@ -3506,59 +3156,23 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/mcp/channel-server.ts: createOpenClawChannelMcpServer",
   "src/mcp/channel-server.ts: OpenClawChannelBridge",
   "src/mcp/openclaw-tools-serve-config.ts: OPENCLAW_TOOLS_MCP_CRESTODIAN_SURFACE_ENV",
-  "src/mcp/openclaw-tools-serve-config.ts: OPENCLAW_TOOLS_MCP_TOOL_IDS",
   "src/mcp/openclaw-tools-serve-config.ts: OpenClawToolsMcpToolId",
   "src/mcp/openclaw-tools-serve-config.ts: resolveOpenClawToolsMcpCrestodianApproval",
   "src/mcp/openclaw-tools-serve-config.ts: resolveOpenClawToolsMcpCrestodianSurface",
   "src/mcp/openclaw-tools-serve-config.ts: resolveOpenClawToolsMcpToolSelection",
   "src/media-understanding/apply.ts: sanitizeMimeType",
-  "src/media-understanding/local-audio.ts: LocalAudioCandidate",
   "src/media-understanding/local-audio.ts: LocalAudioSelection",
   "src/media-understanding/runner.entries.ts: formatMissingProviderHint",
-  "src/media-understanding/runtime-types.ts: MediaUnderstandingScopeContext",
-  "src/media-understanding/types.ts: MediaUnderstandingDocumentModelDefaults",
-  "src/media/image-ops.ts: ResizeToJpegParams",
-  "src/media/image-ops.ts: ResizeToPngParams",
   "src/media/input-files.ts: fetchWithGuard",
-  "src/media/parse.ts: MEDIA_TOKEN_RE",
-  "src/media/parse.ts: normalizeMediaSource",
   "src/media/parse.ts: SplitMediaFromOutputOptions",
-  "src/media/read-capability.ts: createAgentScopedHostMediaReadFile",
-  "src/memory-host-sdk/engine-qmd.ts: QmdBinaryAvailability",
-  "src/memory-host-sdk/engine-qmd.ts: QmdBinaryUnavailable",
-  "src/memory-host-sdk/engine-qmd.ts: QmdBinaryUnavailableReason",
   "src/memory/root-memory-files.ts: resolveCanonicalRootMemoryFile",
   "src/memory/root-memory-files.ts: shouldSkipRootMemoryAuxiliaryPath",
-  "src/model-catalog/authority.ts: mergeModelCatalogRowsByAuthority",
-  "src/model-catalog/index.ts: mergeModelCatalogRowsByAuthority",
-  "src/model-catalog/index.ts: ModelCatalog",
-  "src/model-catalog/index.ts: ModelCatalogAlias",
-  "src/model-catalog/index.ts: ModelCatalogCost",
-  "src/model-catalog/index.ts: ModelCatalogDiscovery",
-  "src/model-catalog/index.ts: ModelCatalogInput",
-  "src/model-catalog/index.ts: ModelCatalogModel",
-  "src/model-catalog/index.ts: ModelCatalogProvider",
-  "src/model-catalog/index.ts: ModelCatalogSource",
-  "src/model-catalog/index.ts: ModelCatalogStatus",
-  "src/model-catalog/index.ts: ModelCatalogSuppression",
-  "src/model-catalog/index.ts: ModelCatalogTieredCost",
-  "src/model-catalog/index.ts: NormalizedModelCatalogRow",
-  "src/model-catalog/index.ts: UnifiedModelCatalogEntry",
-  "src/model-catalog/index.ts: UnifiedModelCatalogKind",
-  "src/model-catalog/index.ts: UnifiedModelCatalogSource",
-  "src/model-catalog/provider-index/index.ts: normalizeOpenClawProviderIndex",
-  "src/model-catalog/provider-index/index.ts: OpenClawProviderIndexPlugin",
-  "src/model-catalog/provider-index/index.ts: OpenClawProviderIndexPluginInstall",
-  "src/model-catalog/provider-index/index.ts: OpenClawProviderIndexProviderAuthChoice",
   "src/music-generation/capabilities.ts: resolveMusicGenerationMode",
   "src/music-generation/runtime-types.ts: ListRuntimeMusicGenerationProvidersParams",
   "src/music-generation/runtime-types.ts: RuntimeMusicGenerationProvider",
-  "src/music-generation/runtime.ts: GenerateMusicParams",
-  "src/music-generation/runtime.ts: GenerateMusicRuntimeResult",
   "src/music-generation/runtime.ts: MusicGenerationRuntimeDeps",
   "src/node-host/exec-policy.ts: formatSystemRunAllowlistMissMessage",
   "src/node-host/invoke-system-run.ts: HandleSystemRunInvokeOptions",
-  "src/node-host/invoke.ts: boundMcpToolResultPayload",
   "src/node-host/invoke.ts: buildNodeEventParams",
   "src/node-host/invoke.ts: decodeCapturedOutputBuffer",
   "src/node-host/invoke.ts: sanitizeEnv",
@@ -3566,38 +3180,18 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/node-host/mcp.ts: buildNodeMcpToolDescriptors",
   "src/node-host/mcp.ts: countConfiguredNodeHostMcpServers",
   "src/node-host/mcp.ts: NodeHostMcpErrorCode",
-  "src/node-host/runner.ts: buildNodeEventParams",
   "src/node-host/runner.ts: buildNodeInvokeResultParams",
   "src/node-host/runner.ts: handleNodeHostReconnectPaused",
   "src/node-host/runner.ts: resolveNodeHostGatewayCredentials",
   "src/node-host/runner.ts: resolveNodeHostGatewayDeviceFamily",
   "src/node-host/runner.ts: resolveNodeHostGatewayPlatform",
   "src/node-host/runner.ts: shouldExitNodeHostOnReconnectPaused",
-  "src/pairing/pairing-store-sqlite.ts: ChannelPairingState",
-  "src/pairing/pairing-store-sqlite.ts: readChannelPairingStateSnapshot",
-  "src/pairing/pairing-store-sqlite.ts: writeChannelPairingStateSnapshot",
   "src/pairing/pairing-store.types.ts: ReadChannelAllowFromStoreForAccount",
   "src/pairing/pairing-store.types.ts: UpsertChannelPairingRequestForAccount",
-  "src/pairing/setup-code.ts: PairingSetupCommandResult",
-  "src/pairing/setup-code.ts: PairingSetupCommandRunner",
-  "src/pairing/setup-code.ts: PairingSetupPayload",
-  "src/pairing/setup-code.ts: PairingSetupResolution",
-  "src/pairing/setup-code.ts: ResolvePairingSetupOptions",
   "src/plugin-state/plugin-state-store.sqlite.ts: probePluginStateStore",
   "src/plugin-state/plugin-state-store.sqlite.ts: seedPluginStateDatabaseEntriesForTests",
   "src/plugin-state/plugin-state-store.sqlite.ts: setMaxPluginStateEntriesPerPluginForTests",
   "src/plugin-state/plugin-state-store.ts: clearPluginStateStoreForTests",
-  "src/plugin-state/plugin-state-store.ts: PluginStateOverflowPolicy",
-  "src/plugin-state/plugin-state-store.ts: PluginStateStoreError",
-  "src/plugin-state/plugin-state-store.ts: PluginStateStoreErrorCode",
-  "src/plugin-state/plugin-state-store.ts: PluginStateStoreOperation",
-  "src/plugin-state/plugin-state-store.ts: PluginStateStoreProbeResult",
-  "src/plugin-state/plugin-state-store.ts: PluginStateStoreProbeStep",
-  "src/plugin-state/plugin-state-store.ts: probePluginStateStore",
-  "src/plugin-state/plugin-state-store.ts: setMaxPluginStateEntriesPerPluginForTests",
-  "src/plugin-state/plugin-state-store.types.ts: PluginStateStoreErrorOptions",
-  "src/plugin-state/runtime-health-store.ts: RuntimeHealthStore",
-  "src/plugin-state/runtime-health-store.ts: RuntimeHealthStoreOptions",
   "src/plugins/activation-context.ts: BundledPluginCompatibleActivationInputs",
   "src/plugins/activation-context.ts: BundledPluginCompatibleLoadValues",
   "src/plugins/activation-context.ts: PluginActivationCompatConfig",
@@ -4141,184 +3735,45 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/plugins/web-provider-types.ts: WebFetchProviderConfiguredCredentialFallback",
   "src/plugins/web-provider-types.ts: WebSearchProviderConfiguredCredentialFallback",
   "src/plugins/worker-provider-registry.ts: listWorkerProviderRegistrations",
-  "src/process/child-process-bridge.ts: ChildProcessBridgeOptions",
   "src/process/command-queue.ts: CommandLaneTaskTimeoutError",
   "src/process/command-queue.ts: resetCommandQueueStateForTest",
-  "src/process/gateway-work-admission.ts: GatewayRootWorkAdmissionLease",
-  "src/process/gateway-work-admission.ts: GatewaySuspendAdmissionLease",
-  "src/process/gateway-work-admission.ts: GatewaySuspendAdmissionPhase",
-  "src/process/gateway-work-admission.ts: runWithGatewayRootWorkAdmission",
   "src/process/gateway-work-admission.ts: tryBeginGatewayIndependentRootWorkAdmission",
-  "src/process/linux-oom-score.ts: hardenedEnvForChildOomWrap",
-  "src/process/linux-oom-score.ts: wrapArgvForChildOomScoreRaise",
-  "src/process/supervisor/adapters/child.ts: ChildAdapter",
-  "src/process/supervisor/adapters/pty.ts: PtyAdapter",
-  "src/process/supervisor/index.ts: createProcessSupervisor",
-  "src/process/supervisor/index.ts: RunExit",
-  "src/process/supervisor/index.ts: RunRecord",
-  "src/process/supervisor/index.ts: RunState",
-  "src/process/supervisor/index.ts: SpawnInput",
-  "src/process/supervisor/index.ts: TerminationReason",
-  "src/proxy-capture/env.ts: OPENCLAW_DEBUG_PROXY_BLOB_DIR",
-  "src/proxy-capture/env.ts: OPENCLAW_DEBUG_PROXY_CERT_DIR",
-  "src/proxy-capture/env.ts: OPENCLAW_DEBUG_PROXY_DB_PATH",
   "src/proxy-capture/env.ts: OPENCLAW_DEBUG_PROXY_ENABLED",
-  "src/proxy-capture/env.ts: OPENCLAW_DEBUG_PROXY_REQUIRE",
   "src/proxy-capture/env.ts: OPENCLAW_DEBUG_PROXY_SESSION_ID",
-  "src/proxy-capture/env.ts: OPENCLAW_DEBUG_PROXY_URL",
   "src/proxy-capture/proxy-server.ts: assertDebugProxyDirectUpstreamAllowed",
   "src/proxy-capture/proxy-server.ts: parseConnectTarget",
-  "src/proxy-capture/store.sqlite.ts: DebugProxyCaptureStoreOptions",
-  "src/proxy-capture/store.sqlite.ts: LegacyDebugProxyCaptureStore",
-  "src/proxy-capture/store.sqlite.ts: SharedDebugProxyCaptureStore",
   "src/realtime-transcription/websocket-session.ts: REALTIME_TRANSCRIPTION_WS_MAX_PAYLOAD_BYTES",
-  "src/secrets/audit.ts: SecretsAuditCode",
-  "src/secrets/audit.ts: SecretsAuditFinding",
-  "src/secrets/audit.ts: SecretsAuditReport",
-  "src/secrets/audit.ts: SecretsAuditSeverity",
-  "src/secrets/audit.ts: SecretsAuditStatus",
-  "src/secrets/auth-profiles-scan.ts: AuthProfileCredentialType",
-  "src/secrets/auth-profiles-scan.ts: AuthProfileCredentialVisit",
-  "src/secrets/channel-contract-api.ts: BundledChannelSecretContractApi",
-  "src/secrets/channel-contract-api.ts: BundledChannelSecurityContractApi",
   "src/secrets/channel-contract-api.ts: loadBundledChannelSecretContractApi",
   "src/secrets/channel-contract-api.ts: loadBundledChannelSecurityContractApi",
-  "src/secrets/channel-env-vars.ts: isSafeChannelEnvVarTriggerName",
-  "src/secrets/command-config.ts: AnalyzeAssignmentsFromSnapshotResult",
-  "src/secrets/command-config.ts: collectCommandSecretAssignmentsFromSnapshot",
-  "src/secrets/command-config.ts: ResolveAssignmentsFromSnapshotResult",
-  "src/secrets/configure-plan.ts: buildConfigureCandidates",
-  "src/secrets/configure-plan.ts: ConfigureProviderChanges",
-  "src/secrets/configure-plan.ts: ConfigureSelectedTarget",
-  "src/secrets/configure.ts: SecretsConfigureResult",
-  "src/secrets/legacy-secretref-env-marker.ts: LegacySecretRefEnvMarkerCandidate",
-  "src/secrets/plan.ts: SecretsPlanTargetType",
-  "src/secrets/provider-env-vars.ts: __testing",
   "src/secrets/provider-env-vars.ts: PROVIDER_AUTH_ENV_VAR_CANDIDATES",
   "src/secrets/provider-env-vars.ts: PROVIDER_ENV_VARS",
   "src/secrets/provider-env-vars.ts: resolveProviderAuthEvidence",
   "src/secrets/provider-env-vars.ts: testing",
-  "src/secrets/ref-contract.ts: EXEC_SECRET_REF_ID_JSON_SCHEMA_PATTERN",
-  "src/secrets/ref-contract.ts: ExecSecretRefIdValidationReason",
-  "src/secrets/ref-contract.ts: ExecSecretRefIdValidationResult",
-  "src/secrets/ref-contract.ts: FILE_SECRET_REF_ID_ABSOLUTE_JSON_SCHEMA_PATTERN",
-  "src/secrets/ref-contract.ts: FILE_SECRET_REF_ID_INVALID_ESCAPE_JSON_SCHEMA_PATTERN",
-  "src/secrets/ref-contract.ts: SecretRefDefaultsCarrier",
-  "src/secrets/runtime-command-secrets.ts: CommandSecretProviderOverrides",
-  "src/secrets/runtime-gateway-auth-surfaces.ts: GatewayAuthSurfacePath",
-  "src/secrets/runtime-gateway-auth-surfaces.ts: GatewayAuthSurfaceState",
-  "src/secrets/runtime-gateway-auth-surfaces.ts: GatewayAuthSurfaceStateMap",
-  "src/secrets/runtime-web-tools.shared.ts: ResolveRuntimeWebProviderSurfaceParams",
-  "src/secrets/runtime-web-tools.shared.ts: RuntimeWebProviderMetadataBase",
-  "src/secrets/runtime-web-tools.shared.ts: RuntimeWebProviderSelectionParams",
-  "src/secrets/runtime-web-tools.shared.ts: RuntimeWebProviderSurface",
-  "src/secrets/runtime-web-tools.ts: RuntimeWebDiagnostic",
-  "src/secrets/runtime-web-tools.ts: RuntimeWebDiagnosticCode",
-  "src/secrets/runtime-web-tools.ts: RuntimeWebFetchMetadata",
-  "src/secrets/runtime-web-tools.ts: RuntimeWebSearchMetadata",
-  "src/secrets/runtime-web-tools.ts: RuntimeWebToolsMetadata",
-  "src/secrets/secret-value.ts: SecretExpectedResolvedValue",
-  "src/secrets/storage-scan.ts: ReadJsonObjectOptions",
   "src/secrets/target-registry-data.ts: getSourceSecretTargetRegistry",
-  "src/secrets/target-registry-pattern.ts: ExpandedPathMatch",
   "src/secrets/target-registry-pattern.ts: parsePathPattern",
-  "src/secrets/target-registry-pattern.ts: PathPatternToken",
-  "src/secrets/target-registry-query.ts: AuthProfileType",
-  "src/secrets/target-registry-query.ts: SecretTargetConfigFile",
-  "src/secrets/target-registry-query.ts: SecretTargetExpected",
-  "src/secrets/target-registry-query.ts: SecretTargetRegistryEntry",
-  "src/secrets/target-registry-query.ts: SecretTargetShape",
-  "src/secrets/target-registry-types.ts: AuthProfileType",
-  "src/secrets/target-registry-types.ts: SecretTargetExpected",
-  "src/secrets/target-registry-types.ts: SecretTargetShape",
   "src/secrets/unsupported-surface-policy.ts: getUnsupportedSecretRefSurfacePatterns",
-  "src/secrets/unsupported-surface-policy.ts: UnsupportedSecretRefConfigCandidate",
-  "src/security/audit-extra.async.ts: collectPluginsTrustFindings",
-  "src/security/audit-extra.async.ts: SecurityAuditFinding",
-  "src/security/audit-extra.summary.ts: SecurityAuditFinding",
-  "src/security/audit-extra.sync.ts: GatewayHttpNoAuthAuditOptions",
-  "src/security/audit-extra.sync.ts: HooksHardeningAuditOptions",
-  "src/security/audit-extra.sync.ts: SecurityAuditFinding",
-  "src/security/audit-gateway-config.ts: CollectGatewayConfigFindingsOptions",
-  "src/security/audit-model-refs.ts: AuditModelRef",
-  "src/security/audit.ts: AuditExecutionContext",
   "src/security/audit.ts: collectElevatedFindings",
   "src/security/audit.ts: collectExecRuntimeFindings",
   "src/security/audit.ts: collectFilesystemFindings",
   "src/security/audit.ts: collectGatewayConfigFindings",
   "src/security/audit.ts: collectLoggingFindings",
   "src/security/audit.ts: collectPluginSecurityAuditFindings",
-  "src/security/audit.ts: SecurityAuditFinding",
-  "src/security/audit.ts: SecurityAuditOptions",
-  "src/security/audit.ts: SecurityAuditSeverity",
-  "src/security/audit.ts: SecurityAuditSummary",
-  "src/security/config-regex.ts: CompiledConfigRegex",
-  "src/security/dangerous-config-flags-core.ts: DangerousConfigFlagContractInputs",
-  "src/security/exec-filesystem-policy.ts: ExecFilesystemPolicyDriftHit",
   "src/security/fix.ts: applySecurityFixConfigMutations",
   "src/security/fix.ts: collectSecurityPermissionTargets",
-  "src/security/fix.ts: SecurityFixAction",
-  "src/security/fix.ts: SecurityFixChmodAction",
-  "src/security/fix.ts: SecurityFixIcaclsAction",
-  "src/security/fix.ts: SecurityFixResult",
-  "src/security/fix.ts: SecurityPermissionTarget",
   "src/security/install-policy.ts: InstallPolicyRequest",
-  "src/security/install-policy.ts: InstallPolicyResult",
-  "src/security/install-policy.ts: InstallPolicyTarget",
-  "src/security/install-policy.ts: InstallPolicyValidationIssue",
-  "src/security/windows-acl.ts: formatWindowsAclSummary",
-  "src/security/windows-acl.ts: inspectWindowsAcl",
-  "src/security/windows-acl.ts: parseIcaclsOutput",
-  "src/security/windows-acl.ts: resolveWindowsUserPrincipal",
-  "src/security/windows-acl.ts: summarizeWindowsAcl",
-  "src/security/windows-acl.ts: WindowsAclEntry",
-  "src/security/windows-acl.ts: WindowsAclSummary",
-  "src/sessions/agent-harness-session-key.ts: AGENT_HARNESS_SESSION_KEY_PREFIX",
-  "src/sessions/input-provenance.ts: INPUT_PROVENANCE_KIND_VALUES",
-  "src/sessions/input-provenance.ts: InputProvenanceKind",
-  "src/sessions/model-overrides.ts: assertModelSelectionUnlocked",
-  "src/sessions/model-overrides.ts: ModelOverrideSelection",
-  "src/sessions/session-chat-type-shared.ts: CanonicalSessionPeerShape",
-  "src/sessions/session-chat-type.ts: deriveSessionChatTypeFromKey",
-  "src/sessions/session-chat-type.ts: SessionKeyChatType",
   "src/sessions/session-key-utils.ts: isCasePreservingPeer",
-  "src/sessions/session-key-utils.ts: ParsedCronRunScopeSuffix",
-  "src/sessions/session-key-utils.ts: ParsedSessionDeliveryRoute",
   "src/sessions/session-lifecycle-admission.ts: runExclusiveSessionLifecycle",
-  "src/sessions/session-lifecycle-admission.ts: SessionLifecycleMutationKind",
   "src/sessions/session-state-events.ts: pruneSessionStateEvents",
   "src/sessions/session-state-events.ts: recordSessionStateEvent",
-  "src/sessions/session-state-events.ts: SessionStateEventInput",
-  "src/sessions/session-state-events.ts: SessionStateEventKind",
-  "src/sessions/session-state-events.ts: SessionStateEventRecord",
   "src/sessions/session-state-events.ts: sessionStateEventStoreLimits",
-  "src/sessions/transcript-events.ts: emitInternalSessionTranscriptUpdate",
-  "src/sessions/transcript-events.ts: SessionTranscriptUpdateTarget",
-  "src/sessions/user-turn-transcript-runtime-context.ts: RuntimeUserTurnTranscriptContext",
-  "src/sessions/user-turn-transcript.ts: attachRuntimeUserTurnTranscriptContext",
-  "src/sessions/user-turn-transcript.ts: attachRuntimeUserTurnTranscriptRecorder",
   "src/sessions/user-turn-transcript.ts: persistUserTurnTranscript",
-  "src/sessions/user-turn-transcript.ts: RuntimeUserTurnTranscriptContext",
-  "src/sessions/user-turn-transcript.ts: takeRuntimeUserTurnTranscriptContext",
-  "src/sessions/user-turn-transcript.ts: takeRuntimeUserTurnTranscriptRecorder",
-  "src/sessions/user-turn-transcript.types.ts: UserTurnTranscriptPersistenceTarget",
-  "src/shared/avatar-policy.ts: AVATAR_MAX_DATA_URL_CHARS",
-  "src/shared/avatar-policy.ts: isAvatarImageDataUrl",
-  "src/shared/avatar-policy.ts: isWorkspaceRelativeAvatarPath",
   "src/shared/config-eval.ts: evaluateRuntimeRequires",
   "src/shared/config-eval.ts: isTruthy",
   "src/shared/config-eval.ts: resolveConfigPath",
   "src/shared/config-eval.ts: resolveRuntimePlatform",
-  "src/shared/custom-command-config.ts: CustomCommandConfig",
-  "src/shared/custom-command-config.ts: CustomCommandInput",
-  "src/shared/custom-command-config.ts: CustomCommandIssue",
   "src/shared/device-auth-store.ts: DeviceAuthStoreAdapter",
-  "src/shared/device-pairing-access.ts: PendingDeviceApprovalState",
   "src/shared/entry-status.ts: evaluateEntryMetadataRequirements",
   "src/shared/entry-status.ts: evaluateEntryMetadataRequirementsForCurrentPlatform",
-  "src/shared/frontmatter.ts: OpenClawManifestRequires",
-  "src/shared/frontmatter.ts: ParsedOpenClawManifestInstallBase",
-  "src/shared/lazy-runtime.ts: LazyPromiseLoader",
   "src/shared/node-match.ts: normalizeNodeKey",
   "src/shared/requirements.ts: buildConfigChecks",
   "src/shared/requirements.ts: evaluateRequirements",
@@ -4328,10 +3783,7 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/shared/requirements.ts: resolveMissingEnv",
   "src/shared/requirements.ts: resolveMissingOs",
   "src/shared/runtime-import.ts: resolveRuntimeImportSpecifier",
-  "src/shared/session-types.ts: GatewayAgentIdentity",
-  "src/shared/session-types.ts: GatewayAgentModel",
   "src/shared/silent-reply-policy.ts: DEFAULT_SILENT_REPLY_POLICY",
-  "src/shared/store-writer-queue.ts: StoreWriterTask",
   "src/shared/subagents-format.ts: formatTokenShort",
   "src/shared/subagents-format.ts: resolveIoTokens",
   "src/skills/discovery/chat-commands.ts: testing",
@@ -4340,26 +3792,15 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/skills/discovery/skill-index.ts: isSkillRuntimeVisible",
   "src/skills/discovery/skill-index.ts: isSkillUserInvocable",
   "src/skills/lifecycle/gh-config-discovery.ts: GhConfigDirMismatch",
-  "src/skills/lifecycle/install.ts: SkillInstallRequest",
-  "src/skills/lifecycle/install.ts: SkillInstallResult",
   "src/skills/lifecycle/install.ts: testing",
   "src/skills/lifecycle/upload-store.ts: createSkillUploadStore",
   "src/skills/lifecycle/upload-store.ts: MAX_ACTIVE_SKILL_UPLOADS",
   "src/skills/loading/plugin-skills.ts: testing",
-  "src/skills/research/signals.ts: extractDurableInstructionProposals",
-  "src/skills/research/signals.ts: WorkspaceSkillSummary",
   "src/skills/runtime/refresh.ts: bumpSkillsSnapshotVersion",
-  "src/skills/runtime/refresh.ts: getSkillsSnapshotVersion",
   "src/skills/runtime/refresh.ts: resetSkillsRefreshForTest",
   "src/skills/runtime/refresh.ts: shouldIgnoreSkillsWatchPath",
-  "src/skills/runtime/refresh.ts: shouldRefreshSnapshotForVersion",
-  "src/skills/runtime/refresh.ts: SkillsChangeEvent",
   "src/skills/runtime/remote-skills.ts: resetRemoteNodeSkillsForTests",
   "src/skills/runtime/session-snapshot.ts: resetResolvedSkillsCacheForTests",
-  "src/skills/runtime/session-snapshot.ts: ReusableSkillSnapshotParams",
-  "src/skills/runtime/session-snapshot.ts: ReusableSkillSnapshotResult",
-  "src/skills/types.ts: SkillCommandDispatchSpec",
-  "src/skills/types.ts: SkillExposure",
   "src/skills/workshop/curator.ts: ARCHIVE_AFTER_MS",
   "src/skills/workshop/curator.ts: CURATOR_INITIAL_DELAY_MS",
   "src/skills/workshop/curator.ts: CURATOR_SWEEP_INTERVAL_MS",
@@ -4367,126 +3808,47 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/skills/workshop/curator.ts: recordSkillUsage",
   "src/skills/workshop/curator.ts: registerSkillUsageTracking",
   "src/skills/workshop/curator.ts: runSkillCuratorSweep",
-  "src/skills/workshop/curator.ts: SkillCuratorSweepResult",
-  "src/skills/workshop/curator.ts: SkillLifecycleState",
-  "src/skills/workshop/curator.ts: SkillOverlapCandidate",
   "src/skills/workshop/curator.ts: STALE_AFTER_MS",
-  "src/skills/workshop/service.ts: WritableWorkspaceSkillSummary",
-  "src/skills/workshop/tool-policy-diagnostic.ts: SkillWorkshopToolPolicyAvailability",
-  "src/skills/workshop/tool-policy-diagnostic.ts: SkillWorkshopToolPolicyDiagnostic",
-  "src/state/openclaw-agent-db.paths.ts: OpenClawAgentSqlitePathOptions",
-  "src/status/codex-synthetic-usage.ts: CODEX_SYNTHETIC_USAGE_HOOK_PROVIDER",
-  "src/status/codex-synthetic-usage.ts: CODEX_SYNTHETIC_USAGE_PROVIDER",
-  "src/status/status-message.ts: StatusArgs",
   "src/status/status-text.ts: resolveStatusChannelFeatureLine",
-  "src/talk/agent-consult-runtime.ts: RealtimeVoiceAgentConsultContextMode",
-  "src/talk/agent-consult-runtime.ts: resolveRealtimeVoiceAgentConsultTools",
-  "src/talk/agent-consult-runtime.ts: resolveRealtimeVoiceAgentConsultToolsAllow",
   "src/talk/agent-consult-runtime.ts: setRealtimeVoiceAgentConsultDepsForTest",
-  "src/talk/agent-run-control.ts: RealtimeVoiceAgentRunActivity",
-  "src/talk/provider-types.ts: RealtimeVoiceBrowserGatewayRelaySession",
-  "src/talk/provider-types.ts: RealtimeVoiceBrowserJsonPcmWebSocketSession",
-  "src/talk/provider-types.ts: RealtimeVoiceBrowserManagedRoomSession",
-  "src/talk/provider-types.ts: RealtimeVoiceBrowserWebRtcSdpSession",
-  "src/talk/talk-session-controller.ts: TalkEventContext",
   "src/tasks/cron-task-cancel.ts: CRON_TASK_RUN_SETTLEMENT_TRACKING_MAX_MS",
   "src/tasks/cron-task-cancel.ts: resetActiveCronTaskRunsForTests",
-  "src/tasks/detached-task-runtime-contract.ts: DetachedTaskCancelParams",
-  "src/tasks/detached-task-runtime-contract.ts: DetachedTaskCancelResult",
-  "src/tasks/detached-task-runtime-contract.ts: DetachedTaskCompleteParams",
-  "src/tasks/detached-task-runtime-contract.ts: DetachedTaskDeliveryStatusParams",
-  "src/tasks/detached-task-runtime-contract.ts: DetachedTaskFailParams",
-  "src/tasks/detached-task-runtime-contract.ts: DetachedTaskProgressParams",
-  "src/tasks/detached-task-runtime-contract.ts: DetachedTaskStartParams",
   "src/tasks/detached-task-runtime-state.ts: DetachedTaskLifecycleRuntime",
-  "src/tasks/detached-task-runtime-state.ts: DetachedTaskLifecycleRuntimeRegistration",
   "src/tasks/detached-task-runtime.ts: cancelDetachedTaskRunById",
-  "src/tasks/detached-task-runtime.ts: DetachedTaskLifecycleRuntime",
-  "src/tasks/detached-task-runtime.ts: DetachedTaskLifecycleRuntimeRegistration",
   "src/tasks/detached-task-runtime.ts: getDetachedTaskLifecycleRuntimeRegistration",
   "src/tasks/detached-task-runtime.ts: registerDetachedTaskRuntime",
   "src/tasks/detached-task-runtime.ts: resetDetachedTaskLifecycleRuntimeForTests",
   "src/tasks/detached-task-runtime.ts: setDetachedTaskLifecycleRuntime",
   "src/tasks/generated-media-task-activity.ts: resetGeneratedMediaTaskActivityForTests",
-  "src/tasks/native-subagent-task.ts: COPILOT_NATIVE_SUBAGENT_RUN_ID_PREFIX",
-  "src/tasks/native-subagent-task.ts: COPILOT_NATIVE_SUBAGENT_TASK_KIND",
-  "src/tasks/native-subagent-task.ts: NativeSubagentTaskDefinition",
-  "src/tasks/runtime-internal.ts: findLatestTaskForFlowId",
-  "src/tasks/runtime-internal.ts: reloadTaskRegistryFromStore",
-  "src/tasks/runtime-internal.ts: resetTaskRegistryControlRuntimeForTests",
-  "src/tasks/runtime-internal.ts: resetTaskRegistryDeliveryRuntimeForTests",
   "src/tasks/runtime-internal.ts: resetTaskRegistryForTests",
-  "src/tasks/runtime-internal.ts: setTaskRegistryControlRuntimeForTests",
-  "src/tasks/runtime-internal.ts: setTaskRegistryDeliveryRuntimeForTests",
   "src/tasks/task-executor.ts: runTaskInFlow",
   "src/tasks/task-flow-registry.store.ts: configureTaskFlowRegistryRuntime",
-  "src/tasks/task-flow-registry.store.ts: TaskFlowRegistryObservers",
-  "src/tasks/task-flow-registry.store.ts: TaskFlowRegistryStore",
-  "src/tasks/task-flow-registry.store.ts: TaskFlowRegistryStoreSnapshot",
   "src/tasks/task-flow-registry.ts: createFlowRecord",
-  "src/tasks/task-flow-registry.ts: CreateFlowRecordParams",
   "src/tasks/task-flow-registry.ts: resetTaskFlowRegistryForTests",
   "src/tasks/task-flow-registry.ts: syncFlowFromTask",
-  "src/tasks/task-flow-registry.ts: TaskFlowSyncResult",
   "src/tasks/task-flow-runtime-internal.ts: resetTaskFlowRegistryForTests",
-  "src/tasks/task-flow-runtime-internal.ts: syncFlowFromTask",
-  "src/tasks/task-flow-runtime-internal.ts: TaskFlowSyncResult",
-  "src/tasks/task-registry-control.types.ts: CancelAcpSessionAdmin",
-  "src/tasks/task-registry-control.types.ts: KillSubagentRunAdmin",
-  "src/tasks/task-registry-control.types.ts: KillSubagentRunAdminResult",
-  "src/tasks/task-registry.audit.ts: createEmptyTaskAuditSummary",
-  "src/tasks/task-registry.audit.ts: TaskAuditOptions",
   "src/tasks/task-registry.ts: findLatestTaskForFlowId",
   "src/tasks/task-registry.ts: findLatestTaskForRelatedSessionKey",
   "src/tasks/task-registry.ts: maybeDeliverTaskStateChangeUpdate",
-  "src/tasks/task-registry.ts: ParentFlowLinkErrorCode",
   "src/tasks/task-registry.ts: resetTaskRegistryControlRuntimeForTests",
   "src/tasks/task-registry.ts: resetTaskRegistryDeliveryRuntimeForTests",
   "src/tasks/task-registry.ts: resetTaskRegistryForTests",
   "src/tasks/task-registry.ts: setTaskRegistryControlRuntimeForTests",
   "src/tasks/task-registry.ts: setTaskRegistryDeliveryRuntimeForTests",
-  "src/tools/diagnostics.ts: ToolPlanContractErrorCode",
-  "src/tools/protocol.ts: ToolProtocolDescriptor",
-  "src/trajectory/paths.ts: resolveTrajectoryPointerOpenFlags",
-  "src/trajectory/runtime-store.sqlite.ts: loadSqliteTrajectoryRuntimeEventsSync",
-  "src/trajectory/runtime-store.sqlite.ts: readSqliteTrajectoryRuntimeStatsSync",
-  "src/trajectory/runtime-store.sqlite.ts: SqliteTrajectoryRuntimeEventRow",
-  "src/trajectory/runtime-store.sqlite.ts: SqliteTrajectoryRuntimeStats",
-  "src/tts/directives.ts: TtsDirectiveTextStreamCleaner",
-  "src/tui/commands.ts: ParsedCommand",
-  "src/tui/commands.ts: SlashCommandOptions",
   "src/tui/components/filterable-select-list.ts: FilterableSelectListTheme",
-  "src/tui/gateway-chat.ts: GatewayAgentsList",
-  "src/tui/gateway-chat.ts: GatewayConnectionOptions",
-  "src/tui/gateway-chat.ts: GatewayEvent",
-  "src/tui/gateway-chat.ts: GatewayModelChoice",
-  "src/tui/gateway-chat.ts: GatewaySessionList",
   "src/tui/gateway-chat.ts: resolveBoundGatewayConnection",
-  "src/tui/gateway-chat.ts: ResolvedGatewayConnection",
   "src/tui/gateway-chat.ts: resolveGatewayConnection",
   "src/tui/theme/theme.ts: lightMode",
   "src/tui/theme/theme.ts: lightPalette",
   "src/tui/theme/theme.ts: palette",
-  "src/tui/tui-backend.ts: TuiGoalCommandOptions",
-  "src/tui/tui-backend.ts: TuiTaskSuggestionActionCapabilities",
   "src/tui/tui-busy-notice.ts: TUI_AGENT_BUSY_MESSAGE",
   "src/tui/tui-last-session.ts: isHeartbeatLikeTuiSession",
   "src/tui/tui-last-session.ts: resolveTuiLastSessionStatePath",
   "src/tui/tui-plugin-approvals.ts: parseTuiPluginApproval",
   "src/tui/tui-task-suggestions.ts: parseTuiTaskSuggestion",
-  "src/tui/tui-types.ts: QueuedMessage",
-  "src/tui/tui-types.ts: QueuedMessageMode",
-  "src/tui/tui-types.ts: TuiExitReason",
   "src/tui/tui-waiting.ts: pickWaitingPhrase",
-  "src/utils/delivery-context.shared.ts: DeliveryContextSessionSource",
-  "src/utils/message-channel-constants.ts: InternalMessageChannel",
   "src/utils/message-channel-constants.ts: NATIVE_APPROVAL_CHANNELS",
-  "src/utils/message-channel-constants.ts: NativeApprovalChannel",
-  "src/utils/message-channel.ts: InternalMessageChannel",
-  "src/utils/message-channel.ts: NATIVE_APPROVAL_CHANNELS",
-  "src/utils/message-channel.ts: NativeApprovalChannel",
   "src/utils/queue-helpers.ts: clearQueueSummaryState",
-  "src/utils/timer-delay.ts: resolveFiniteTimeoutDelayMs",
   "src/video-generation/runtime-types.ts: ListRuntimeVideoGenerationProvidersParams",
   "src/video-generation/runtime-types.ts: RuntimeVideoGenerationProvider",
   "src/web-search/runtime-types.ts: ListWebSearchProvidersParams",
@@ -4495,20 +3857,10 @@ export const KNIP_UNUSED_EXPORT_BASELINE = [
   "src/wizard/i18n/index.ts: listWizardI18nKeys",
   "src/wizard/i18n/index.ts: resolveWizardLocale",
   "src/wizard/i18n/index.ts: resolveWizardLocaleFromEnv",
-  "src/wizard/i18n/index.ts: WIZARD_DEFAULT_LOCALE",
   "src/wizard/i18n/index.ts: WIZARD_SUPPORTED_LOCALES",
-  "src/wizard/i18n/index.ts: WizardLocale",
-  "src/wizard/i18n/index.ts: WizardTranslationMap",
-  "src/wizard/prompts.ts: WizardDeviceCodeParams",
-  "src/wizard/session.ts: WizardStepOption",
   "src/wizard/setup.migration-import.ts: inspectSetupMigrationFreshness",
-  "src/wizard/setup.migration-import.ts: SetupMigrationDetection",
-  "src/wizard/setup.migration-import.ts: SetupMigrationOption",
-  "src/wizard/setup.official-plugins.ts: OfficialPluginOnboardingInstallEntry",
   "src/wizard/setup.official-plugins.ts: resolveOfficialPluginOnboardingInstallEntries",
   "src/wizard/setup.official-plugins.ts: testing",
-  "src/wizard/setup.post-install-migration.ts: PostInstallMigrationOptions",
-  "src/wizard/setup.post-install-migration.ts: PostInstallMigrationResult",
   "ui/src/api/gateway.ts: CONTROL_UI_BOOTSTRAP_OPERATOR_SCOPES",
   "ui/src/api/gateway.ts: CONTROL_UI_OPERATOR_SCOPES",
   "ui/src/api/gateway.ts: isNonRecoverableConnectError",
